refactor: reorganize file-based OS files
Ref eng-268
