+++ packages/core/src/external/opensearch/file/file-ingestor.ts
@@ -1,5 +1,5 @@
-import { OpenSearchConfig } from ".";
-import { IndexFields } from "./index-based-on-file";
+import { OpenSearchConfig } from "..";
+import { IndexFields } from "../index-based-on-file";
 
 export type IngestRequest = Omit<IndexFields, "content"> & {
   entryId: string;
@@ -14,7 +14,7 @@ export abstract class OpenSearchFileIngestor {
 
   abstract ingest(req: IngestRequest): Promise<void>;
 
-  isIngestible(file: { contentType?: string; fileName: string }) {
+  isIngestible(file: { contentType?: string | undefined; fileName: string }) {
     const ingestibleTypes = ["xml", "text", "txt", "html", "htm", "ascii"];
     if (file.contentType) {
       return ingestibleTypes.some(
