+++ packages/core/src/external/opensearch/file/file-remover.ts
